requirements.txt
# International Wholesale — Cotação Agent
# Cloud-ready dependencies
# Instalar: pip install -r requirements.txt

streamlit>=1.35
anthropic>=0.40.0
openpyxl>=3.1.0
pandas>=2.1.0
supabase>=2.4.0
python-dotenv>=1.0.0
